fix: broken engine support (#82)
* revert change to engine regex from f5569be7d4625

This broke support for engines, which are detected by the
path to specs _not_ starting with 'spec/' (e.g. 'my_engine/spec/'
instead).

* Add test coverage of engine support

tweaks to ensure relative paths are checked instead of absolute ones, so
running within a project with 'spec/' in the path (e.g. neotest-rspec/)
doesn't get treated as an engine.

* Add ability to turn off engine support

#49 suggests that some people prefer not to have this functionality